Switched to virtual ports
Midi in and out are now on virtual ports, so no need to use the IAC bus on OSX or third parts software on windows. The ports will show up once the application is launched. References to port numbers have been removed from the application, but the names of the ports are still shown on the interface.